moving ActivityApiControllerTestsInto AllReady.UnitTest.Controller folder/namesapce
